(MODULES-2917) Modify spec tests for DISM module
After the modsync update, the spec tests were failing due to the
updated rspec, puppet, facter.

This commit updates the spec tests to be more readable.  In
particular, all functionality of the previously existing spec
tests were preserved.  Dism.rb was refactored to be easier to
test.  Metadata.json was changed after the Puppet rebranding.