(MODULES-2917) Fix DISM failing with single quotes in source param
When DISM type was used with parameter source, the desired
feature would fail to install due to the code adding single
quotes to the path.

This commit modifies dism.rb to remove the single quotes, which
fixes the issue for pathnames with and without spaces.  At the
same time, the spec tests have been updated to check against
source parameter without single quotes.  This commit also
adds spec tests for this scenario.